+++ .gitignore
@@ -1,0 +1,3 @@
+instance/config.py
+.pytest_cache
+requirements2.txt
