+++ .travis.yml
@@ -1,7 +1,16 @@
 language: python
+
 python:
   - "3.6"
+
 install:
   - pip install -r requirements.txt
+  - pip install coveralls codecov
+
 script:
   - python3 -m pytest
+  #- nosetests --with-coverage --cover-package=app test
+
+after_success:
+  - coveralls
+  #- codecov --token=bahaslas
