+++ requirements.txt
@@ -1,57 +1,39 @@
-alembic==0.9.9
-atomicwrites==1.2.1
-attrs==18.2.0
-blinker==1.4
-certifi==2018.8.13
+aniso8601==3.0.2
+certifi==2018.8.24
 chardet==3.0.4
 click==6.7
+coverage==4.5.1
 Flask==1.0.2
-Flask-Dance==0.14.0
-Flask-DebugToolbar==0.10.1
-Flask-Login==0.4.1
-Flask-Migrate==2.1.1
-Flask-OAuth==0.12
-Flask-OAuthlib==0.9.4
-Flask-SQLAlchemy==2.3.2
-Flask-WTF==0.14.2
+Flask-RESTful==0.3.6
 gunicorn==19.9.0
 h11==0.7.0
-httplib2==0.11.3
 idna==2.6
 IMAPClient==2.0.0
 itsdangerous==0.24
 jdcal==1.4
 Jinja2==2.10
-lazy==1.3
-Mako==1.0.7
+jsonschema==2.6.0
+lxml==4.2.5
 MarkupSafe==1.0
-more-itertools==4.3.0
-oauth2==1.9.0.post1
-oauthlib==2.0.7
 openpyxl==2.1.4
-pluggy==0.7.1
-py==1.6.0
+Pillow==5.2.0
+PyAutoGUI==0.9.36
+pycrypto==2.6.1
 PyJWT==1.6.4
 PyMsgBox==1.0.6
 PyPDF2==1.26.0
 PyScreeze==0.1.14
 PySocks==1.6.8
-pytest==3.8.0
-python-dateutil==2.7.2
-python-editor==1.0.3
+python-dateutil==2.7.3
+python-docx==0.8.6
+python-dotenv==0.9.1
+python3-xlib==0.15
 PyTweening==1.0.3
 pytz==2018.5
 requests==2.18.4
-requests-oauthlib==0.8.0
 selenium==3.13.0
 Send2Trash==1.5.0
 six==1.11.0
-SQLAlchemy==1.2.6
-SQLAlchemy-Utils==0.33.2
 twilio==6.14.7
 urllib3==1.22
-URLObject==2.4.3
 Werkzeug==0.14.1
-wincertstore==0.2
-WTForms==2.1
-uwsgi
